+++ etna/datasets/tsdataset.py
@@ -398,6 +398,10 @@ class TSDataset:
             for transform in reversed(self.transforms):
                 tslogger.log(f"Inverse transform {repr(transform)} is applied to dataset")
                 self.df = transform.inverse_transform(self.df)
+                columns_after = set(self.columns.get_level_values("feature"))
+                self._regressors = list(set(self._regressors).union(columns_after))
+                if "target" in self._regressors:
+                    self._regressors.remove("target")
 
     @property
     def segments(self) -> List[str]:
